+++ lib/src/widget/clip/clip.dart
@@ -1,4 +1,5 @@
 export 'clip_path.dart';
 export 'clip_rect.dart';
+export 'jozz_clips.dart';
 export 'squircle.dart';
 export 'superellipse.dart';
