+++ lib/src/widget/clip/squircle.dart
@@ -1,6 +1,6 @@
-import 'dart:math';
-
 import 'package:flutter/widgets.dart';
+
+import 'jozz_clips.dart';
 
 /// Extension to clip any widget into a squircle (rounded square) shape.
 ///
@@ -24,42 +24,14 @@ import 'package:flutter/widgets.dart';
 ///
 extension SquircleWidgetExtensions on Widget {
   /// Clips the widget using a squircle shape.
-  Widget clipSquircle([
+  ClipPath clipSquircle([
     double radiusFactor = 2.5,
     Clip clipBehavior = Clip.antiAlias,
   ]) {
     return ClipPath(
-      clipper: _SquircleClipper(radiusFactor),
+      clipper: JozzSquircleClipper(radiusFactor),
       clipBehavior: clipBehavior,
       child: this,
     );
   }
-}
-
-class _SquircleClipper extends CustomClipper<Path> {
-  final double radiusFactor;
-  _SquircleClipper(this.radiusFactor);
-
-  @override
-  Path getClip(Size size) {
-    final double minSize = min(size.width, size.height);
-    final double cornerRadius = minSize / radiusFactor;
-
-    final path = Path();
-    path.moveTo(0, cornerRadius);
-    path.quadraticBezierTo(0, 0, cornerRadius, 0);
-    path.lineTo(size.width - cornerRadius, 0);
-    path.quadraticBezierTo(size.width, 0, size.width, cornerRadius);
-    path.lineTo(size.width, size.height - cornerRadius);
-    path.quadraticBezierTo(
-        size.width, size.height, size.width - cornerRadius, size.height);
-    path.lineTo(cornerRadius, size.height);
-    path.quadraticBezierTo(0, size.height, 0, size.height - cornerRadius);
-    path.close();
-
-    return path;
-  }
-
-  @override
-  bool shouldReclip(covariant CustomClipper<Path> oldClipper) => false;
 }
