+++ lib/src/widget/clip/superellipse.dart
@@ -1,6 +1,6 @@
-import 'dart:math';
-
 import 'package:flutter/widgets.dart';
+
+import 'jozz_clips.dart';
 
 /// Clips the widget into a superellipse (squircle-like) shape using a
 /// [CustomClipper] based on the superellipse equation.
@@ -29,88 +29,15 @@ import 'package:flutter/widgets.dart';
 /// ```
 ///
 extension SuperellipseWidgetExtensions on Widget {
-  Widget clipSuperellipse([
+  ClipPath clipSuperellipse([
     double n = 4,
     int steps = 64,
     Clip clipBehavior = Clip.antiAlias,
   ]) {
     return ClipPath(
-      clipper: _SuperellipseClipper(n, steps),
+      clipper: JozzSuperellipseClipper(n, steps),
       clipBehavior: clipBehavior,
       child: this,
     );
   }
-}
-
-class _SuperellipseClipper extends CustomClipper<Path> {
-  final double n;
-  final int steps;
-
-  const _SuperellipseClipper([
-    this.n = 4,
-    this.steps = 64,
-  ]);
-
-  // Static cache map
-  static final Map<_SuperellipseCacheKey, Path> _cache = {};
-
-  @override
-  Path getClip(Size size) {
-    final key = _SuperellipseCacheKey(n: n, steps: steps, size: size);
-
-    // Return cached path if exists
-    if (_cache.containsKey(key)) return _cache[key]!;
-
-    // Generate and cache
-    final a = size.width / 2;
-    final b = size.height / 2;
-    final cx = a;
-    final cy = b;
-
-    final path = Path();
-    for (int i = 0; i <= steps; i++) {
-      final theta = (2 * pi * i) / steps;
-      final cosT = cos(theta);
-      final sinT = sin(theta);
-      final x = a * pow(cosT.abs(), 2 / n) * cosT.sign;
-      final y = b * pow(sinT.abs(), 2 / n) * sinT.sign;
-
-      if (i == 0) {
-        path.moveTo(cx + x, cy + y);
-      } else {
-        path.lineTo(cx + x, cy + y);
-      }
-    }
-
-    path.close();
-    _cache[key] = path;
-    return path;
-  }
-
-  @override
-  bool shouldReclip(covariant _SuperellipseClipper oldClipper) =>
-      n != oldClipper.n || steps != oldClipper.steps;
-}
-
-class _SuperellipseCacheKey {
-  final double n;
-  final int steps;
-  final Size size;
-
-  const _SuperellipseCacheKey({
-    required this.n,
-    required this.steps,
-    required this.size,
-  });
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is _SuperellipseCacheKey &&
-          n == other.n &&
-          steps == other.steps &&
-          size == other.size;
-
-  @override
-  int get hashCode => n.hashCode ^ steps.hashCode ^ size.hashCode;
 }
